official_writeup/web-copy/game/requirements.txt
flask
flask-caching
pyopenssl
pytz
cryptography
brotli
